html/browsers/sandboxing/sandbox-navigation-timing-2.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Sandbox Navigation Timing</title>
    <script src=/resources/testharness.js></script>
    <script src=/resources/testharnessreport.js></script>
  </head>
  <body>
    <script>
    const resourcePath =
        location.pathname.substring(0, location.pathname.lastIndexOf('/') + 1) +
        'resources/';
    async_test(t => {
      const iframe = document.createElement('iframe');
      iframe.src = resourcePath + 'post-sandbox-state-to-parent.html';
      iframe.sandbox = '';

      // Navigation should be initiated here. The value of the sandbox attribute
      // should be snapshotted.
      document.body.appendChild(iframe);

      // Manipulating the sandbox attribute at this point should not affect the
      // sandbox flags used for either the initial about:blank document or the
      // final document in the iframe.
      iframe.sandbox = 'allow-same-origin';

      // Before navigation, about:blank should be sandboxed.
      assert_equals(iframe.contentDocument, null,
        "about:blank document should be sandboxed");

      // After navigation, child frame should be sandboxed.
      iframe.onload = t.step_func_done(() => {
        assert_equals(iframe.contentDocument, null,
          "loaded document should be sandboxed");
      });
    }, "modifying sandbox attribute after navigation starts should not affect" +
       " document in iframe.");
    </script>
  </body>
</html>
